+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ### Changed
 
+- Create `AppCatalog` CRs from `Catalog` CRs for compatibility with existing app-operator releases.
 - Prepare helm values to configuration management.
 - Use `Catalog` CRs in `App` controller.
 - Reconcile to `Catalog` CRs instead of `AppCatalog`.